requirements.txt
asgiref==3.5.2
autopep8==1.7.0
certifi==2022.6.15.2
cffi==1.15.1
charset-normalizer==2.1.1
contextlib2==21.6.0
cryptography==38.0.1
Django==4.1.1
django-filter==22.1
django-ninja==0.19.1
django-ninja-extra==0.15.6
django-ninja-jwt==5.2.0
django-silk==5.0.1
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.2.1
gprof2dot==2022.7.29
idna==3.4
injector==0.19.0
Jinja2==3.1.2
jose==1.0.0
MarkupSafe==2.1.1
ninja-schema==0.13.0
pyasn1==0.4.8
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.10.2
PyJWT==2.4.0
python-dateutil==2.8.2
python-jose==3.3.0
pytz==2022.2.1
requests==2.28.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
types-cryptography==3.3.23
typing_extensions==4.3.0
tzdata==2022.2
urllib3==1.26.12
